refactoring, structure
refactoring - 함수와 페이지 분리, structure - 수정 관련사항 추가
