--- Frontend/functions/anime_characterize.py
@@ -1,14 +1,0 @@
-from models.anime_characterize import InputModel, OutputModel
-from utils.page import PageModel
-
-
-def execute(
-    page: PageModel,
-    key: str,
-    model: InputModel,
-) -> OutputModel | None:
-    return page.settings.client.call(
-        function=page.function,
-        input=model,
-        output_model=OutputModel,
-    )
