--- Frontend/functions/identity.py
@@ -1,11 +1,0 @@
-from pydantic import BaseModel
-
-from utils.page import PageModel
-
-
-def execute[InputModel: BaseModel](
-    page: PageModel,
-    key: str,
-    model: InputModel,
-) -> InputModel:
-    return model
